PVB-622: PAR start page - authentication refactoring
* Initial set of changes for Auth, ready to merge in httpclient refactor

* Refactoring the location of Authentication code.

Still lots to do!

* Working through unit tests for authentication

* Refactoring and tests

It appears OneLogin is now giving an "Invalid request" response...

* Fix type of EmailVerified

* Minor updates

* Fixed failing unit tests

* Include a readme for Authentication

* Refactor as both StaffSSO and OneLogin share builder

OneLogin seems to be throwing an error at the moment

* Now, both StaffSSO and OneLogin work

Final work is to fix and complete unit test coverage.

* Completing code coverage unit tests

* Refactoring

The Common project was becoming difficult to manage.
This has mostly been renamed to CQRS specific parts of the IHttpService extracted into a ICqrsService. This allows the Http project to not know about Cqrs.

* Bugfix